Implementación de pruebas para la conversión de fechas y generación de archivos de calendario en el módulo ShiftTradeAV. Se añadieron funciones para manejar fechas en diferentes zonas horarias y se creó un archivo de pruebas para validar la funcionalidad. Además, se estableció la estructura del paquete de utilidades, incluyendo autenticación, gestión de empleados, y solicitudes de cambio de turno, con integración a Supabase para almacenamiento y recuperación de datos.
